Arduino: Match the number game using a 128x64 OLED screen
This contains a sketch for a game named "Match the Number". The OLED
screen is split into two parts - left and right. A number remains
fixed on the left side of the screen, while another numbers keeps
ticking away on the right side of the screen. The objective is to
press the push-button when both the numbers match.

Every correct answer takes the player to the next level, where the
speed with which the number on the right ticks away is greater than
the previous level.

Every wrong answer, and the player loses a life.

The objective is to clear all 10 levels.

Signed-off-by: Gautham R. Shenoy <gautham.social@gmail.com>